+++ src/components/Board/BoardCell/BoardCell.styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { flexContainer, blue30, green50, red50, font, text } from 'remember-ui';
+import { flexContainer, blue30, green50, red50 } from 'remember-ui';
 import { Link } from 'react-router-dom';
 
 import detailIcon from 'assets/images/ico-detail.svg';
@@ -23,27 +23,6 @@ export const Container = styled.div`
   text-align: center;
 
   cursor: ${({ isMainBoard }) => (isMainBoard ? 'pointer' : 'move')};
-`;
-
-export const TextArea = styled.textarea.attrs({
-  type: 'text',
-  maxLength: 30,
-})`
-  ${font({ size: '14px', color: text })};
-
-  background-color: transparent;
-  width: 100%;
-  height: 60%;
-  border: 0px;
-  padding: 5px 10px;
-  text-align: center;
-  outline: none;
-  resize: none;
-  -webkit-appearance: none;
-
-  &:focus {
-    border: 0px;
-  }
 `;
 
 export const Hover = styled.div`
